+++ CHANGELOG.md
@@ -22,6 +22,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   English. This was added by @harderj ✨
 - Added the new Danish linguistic acceptability dataset DaLA. It's marked as
   unofficial for now. This was added by @N-essuno ✨
+- Added the new Danish Word in Context dataset DanWiC. It's marked as unofficial for
+  now. This is based on the COR.SEM lexical resource and tests the ability to
+  distinguish word meanings/senses in context.
 - Added `--max-context-length` and `--vocabulary-size` CLI options (and corresponding
   `max_context_length` and `vocabulary_size` arguments to `Benchmarker.__init__` and
   `Benchmarker.benchmark`) to allow overriding the model metadata values that are